fix json loading issue
