+++ frontend/parceltracker/parceltracker/Models/Data.swift
@@ -11,8 +11,8 @@ import CoreLocation
 import UIKit
 import SwiftUI
 
-let courierData: [Courier] = load("mockCourierData.json")
 let parcelLifespan: [Status] = load("trackingData.json")
+let courierData: [Courier] = load("mockCourierData.json")
 
 func load<T: Decodable>(_ filename: String) -> T {
     let data: Data
